Webhooks + Css search issue

+++ VibeSync.DAL/Repository/QueryRepository/IUserQueryRepository.cs
@@ -6,7 +6,7 @@ namespace VibeSync.DAL.Repository.QueryRepository
     public interface IUserQueryRepository
     {
         IEnumerable<User> GetUsers();
-        User GetUserById(int userId);
+        User GetUserById(long userId);
         User GetUserByEmail(string email);
         User ChecksIfUserIsValid(string email, string password);
         long GetDjByUserId(long userId);
